requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.94.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.6.1
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
loguru==0.6.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pyasn1==0.5.0
pydantic==1.10.6
pytest==7.2.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
watchfiles==0.21.0
websockets==12.0
win32-setctime==1.1.0
